move 'boolean_attribute' to shared
